+++ public/psycheName/public/name/psycheName.css
@@ -1,5 +1,5 @@
 body {
-    font-family: 'Arial', sans-serif;
+    font-family: 'Arial', 'Helvetica', sans-serif;
     background: radial-gradient(circle,black,black); /* Deep space gradient */
     color: white;
     text-align: center;
@@ -7,7 +7,14 @@ body {
     padding: 0;
     overflow-x: hidden;
 }
-
+#animation-container {
+    position: absolute;
+    z-index: 1;  /* Lower z-index for the 3D scene */
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+}
 /* Container */
 .container {
     max-width: 800px;
@@ -52,7 +59,8 @@ h2 {
     margin: 20px auto;
     padding: 10px 20px;
     background: #0077cc;
-    color: 2px solid rgb(0, 187, 255);
+    color: white;
+    border: 2px solid rgb(0, 187, 255);
     text-decoration: none;
     border-radius: 5px;
     text-align: center;
@@ -170,8 +178,8 @@ ul li:last-child {
     padding: 20px;
     overflow: auto;
     overflow-y: auto;
-    display: none;
     position: fixed;
+    z-index: 2;
 }
 
 body::before {
