+++ src/landing/psycheNameViewport.js
@@ -4,61 +4,55 @@
  * This module handles loading the psycheName/psycheName.html content in an iframe
  * that appears on top of the Three.js scene.
  */
-
+import * as THREE from 'three';
 import gsap from 'gsap';
-
+import { GLTFLoader } from 'three/examples/jsm/loaders/GLTFLoader.js';
 // Keep track of the viewport DOM elements
 let viewportContainer = null;
 let iframe = null;
 let closeButton = null;
-
 /**
- * Creates and shows the psycheName viewport with animations.
+ * Creates and shows the name viewport
  */
-export function showPsycheNameViewport() {
+export function showNameViewport() {
     // If viewport already exists, just show it
     if (viewportContainer) {
         viewportContainer.style.display = 'flex';
         return;
     }
-
-    console.log("Creating psycheName viewport");
-
+    console.log("Creating Name viewport");
     // Create container for the viewport
     viewportContainer = document.createElement('div');
-    viewportContainer.id = 'psyche-name-viewport-container';
+    viewportContainer.id = 'name-viewport-container';
     viewportContainer.style.position = 'fixed';
     viewportContainer.style.top = '50%';
     viewportContainer.style.left = '50%';
     viewportContainer.style.transform = 'translate(-50%, -50%)';
-    viewportContainer.style.width = '80%';
-    viewportContainer.style.maxWidth = '1440px';
-    viewportContainer.style.height = '80vh';
+    viewportContainer.style.width = '80%'; // Increased from 90% to 110% (20% wider)
+    viewportContainer.style.maxWidth = '1200px'; // Increased from 1200px to 1440px (20% wider)
+    viewportContainer.style.height = '80vh'; // Increased from 85vh to 95vh (10% higher)
     viewportContainer.style.backgroundColor = 'rgba(0, 0, 0, 0.9)';
-    viewportContainer.style.border = '2px solid #f9a000';
-    viewportContainer.style.borderRadius = '10px';
-    viewportContainer.style.boxShadow = '0 0 20px rgba(249, 160, 0, 0.5)';
-    viewportContainer.style.zIndex = '900';
+    viewportContainer.style.border = '0.5px solid #FFFF';
+    viewportContainer.style.borderRadius = '20px';
+    viewportContainer.style.boxShadow = '0 0 30px rgba(0, 123, 255, 0.5)';
+    viewportContainer.style.zIndex = '1000';
     viewportContainer.style.display = 'flex';
     viewportContainer.style.flexDirection = 'column';
     viewportContainer.style.overflow = 'hidden';
-    
     // Create header with title and close button
     const header = document.createElement('div');
     header.style.display = 'flex';
     header.style.justifyContent = 'space-between';
     header.style.alignItems = 'center';
-    header.style.padding = '10px 15px';
-    header.style.backgroundColor = '#f9a000';
+    header.style.padding = '1px 2px';
+    header.style.backgroundColor = 'rgba(0, 0, 0, 0.3)';
     header.style.color = 'white';
-    header.style.borderTopLeftRadius = '8px';
-    header.style.borderTopRightRadius = '8px';
-    
+    header.style.borderTopLeftRadius = '20px';
+    header.style.borderTopRightRadius = '20px';
     const title = document.createElement('h2');
-    title.textContent = 'The Origin of Psyche\'s Name';
-    title.style.margin = '0';
+//    title.textContent = 'Psyche Name Origin';
+//   title.style.margin = '0';
     title.style.fontSize = '1.2rem';
-    
     closeButton = document.createElement('button');
     closeButton.textContent = '✕';
     closeButton.style.background = 'none';
@@ -68,93 +62,123 @@ export function showPsycheNameViewport() {
     closeButton.style.cursor = 'pointer';
     closeButton.style.padding = '0 5px';
     closeButton.style.lineHeight = '1';
-    
-    header.appendChild(title);
+ //   header.appendChild(title);
     header.appendChild(closeButton);
     viewportContainer.appendChild(header);
-    
-    // Create iframe to load the psycheName content
+    // Create iframe to load the balance.html content
     iframe = document.createElement('iframe');
-    iframe.src = 'public/name/psycheName.html';
+    iframe.src = 'public/psycheName/public/name/psycheName.html';
     iframe.style.width = '100%';
     iframe.style.height = '100%';
     iframe.style.border = 'none';
-    iframe.style.backgroundColor = '#fff';
-    
+    iframe.style.backgroundColor = '#222';
     // Add event listener for iframe load errors
     iframe.onerror = () => {
-        console.error("Failed to load psycheName iframe content");
+        console.error("Failed to load iframe content");
     };
-    
     // Add event listener for iframe load success
     iframe.onload = () => {
-        console.log("PsycheName iframe loaded successfully");
+        console.log("Iframe loaded successfully");
     };
-    
     viewportContainer.appendChild(iframe);
     document.body.appendChild(viewportContainer);
-    
-    // Create opening animations using a timeline
-    const tl = gsap.timeline();
-    tl.from(viewportContainer, {
+    // Add animation for opening
+    gsap.from(viewportContainer, {
         opacity: 0,
         scale: 0.8,
         duration: 0.4,
         ease: "power2.out"
     });
-    tl.from(header, {
-        y: -50,
-        duration: 0.3,
-        ease: "back.out(1.7)"
-    }, "-=0.2");
-    tl.from(iframe, {
-        opacity: 0,
-        y: 20,
-        duration: 0.3,
-        ease: "power2.out"
-    }, "-=0.1");
-    
     // Add event listener for close button
-    closeButton.addEventListener('click', hidePsycheNameViewport);
-    
+    closeButton.addEventListener('click', hideNameViewport);
     // Add event listener for Escape key
     document.addEventListener('keydown', handleKeyDown);
 }
+document.addEventListener("DOMContentLoaded", function () {
+    let contentBox = document.getElementById("content");
+    let popup = document.getElementById("viewportContainer");
+    let closeButton = document.getElementById("closeButton");
 
+    if (!popup || !contentBox || !closeButton) {
+        console.error("Missing elements: viewportContainer, content, or closeButton.");
+        return;
+    }
+    // Content for each section
+    const sections = {
+        introLink: "<h2>Introduction</h2><p>Psyche is a metal-rich asteroid orbiting the Sun between Mars and Jupiter. It was named after the Greek goddess Psyche.</p>",
+        historyLink: "<h2>Historical Background</h2><p>Psyche was discovered in 1852 by Italian astronomer Annibale de Gasparis and is one of the largest asteroids in our solar system.</p>",
+        missionLink: "<h2>NASA's Psyche Mission</h2><p>The Psyche spacecraft, designed by NASA, aims to study the asteroid to understand more about planetary cores and the history of the solar system.</p>",
+        funFactsLink: "<h2>Fun Facts</h2></h2></h2><ul><li>Psyche is primarily composed of nickel and iron, much like Earth's core.</li><li>It is about 226 kilometers (140 miles) in diameter.</li><li>Its name means 'soul' in Greek mythology.</li></ul>"
+    };
+
+    // Show popup when clicking a link
+    document.querySelectorAll("ul li a").forEach(link => {
+        link.addEventListener("click", function () {
+            let sectionKey = this.id;
+            contentBox.innerHTML = sections[sectionKey] || "No content available.";
+            popup.style.display = "block";
+            gsap.from(popup, {
+                opacity: 0,
+                scale: 0.8,
+                duration: 0.4,
+                ease: "power2.out"
+            });
+        });
+    });
+
+    // Close popup
+    closeButton.addEventListener("click", hideNameViewport); 
+
+    // Close popup when clicking outside of it
+    window.addEventListener("click", function (e) {
+        if (e.target === popup) {
+            hideNameViewport();
+        }
+    });
+});
 /**
- * Hides the psycheName viewport with closing animation.
+ * Hides the name viewport
  */
-export function hidePsycheNameViewport() {
+export function hideNameViewport() {
     if (!viewportContainer) return;
+//    viewportContainer.style.button = "-100%";
+//    viewportContainer.style.opacity = "0";
 
+//    setTimeout(() => {
+//        document.body.removeChild(viewportContainer);
+//        viewportContainer = null;
+//        iframe = null;
+//        closeButton = null;
+//    }, 500);
+//    };
+    // Animate closing
     gsap.to(viewportContainer, {
         opacity: 0,
-        scale: 0.8,
+       scale: 0.8,
         duration: 0.3,
         ease: "power2.in",
         onComplete: () => {
-            destroyPsycheNameViewport();
-            document.body.classList.add("overlay-open");
+            viewportContainer.style.display = 'none';
+            // Reset opacity and scale for next time
+            viewportContainer.style.opacity = 1;
+            viewportContainer.style.transform = 'translate(-50%, -50%) scale(1)';
         }
     });
 }
-
-
 /**
- * Handles keydown events for the viewport.
+ * Handles keydown events for the viewport
  */
 function handleKeyDown(e) {
     if (e.key === 'Escape') {
-        hidePsycheNameViewport();
+        hideNameViewport();
     }
 }
-
 /**
- * Removes the viewport completely.
+ * Removes the viewport completely
  */
-export function destroyPsycheNameViewport() {
+export function destroyNameViewport() {
     if (viewportContainer) {
-        closeButton.removeEventListener('click', hidePsycheNameViewport);
+        closeButton.removeEventListener('click', hideNameViewport);
         document.removeEventListener('keydown', handleKeyDown);
         document.body.removeChild(viewportContainer);
         viewportContainer = null;
